chore: move preprocess config out of rollup file
